requirements.txt
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.6
django-grappelli==4.0.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0
pytz==2024.1
scikit-surprise==1.1.4
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
ujson==5.10.0
mysqlclient==2.2.4
FuzzyTM==1.0.0
python-lsp-server==1.10.0
